% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_wrap_paginate.R
\name{facet_wrap_paginate}
\alias{facet_wrap_paginate}
\title{Split facet_wrap over multiple plots}
\usage{
facet_wrap_paginate(facets, nrow = NULL, ncol = NULL, scales = "fixed",
  shrink = TRUE, labeller = "label_value", as.table = TRUE,
  switch = NULL, drop = TRUE, dir = "h", strip.position = "top",
  page = 1)
}
\arguments{
\item{facets}{Either a formula or character vector. Use either a
one sided formula, \code{~a + b}, or a character vector, \code{c("a", "b")}.}

\item{nrow}{Number of rows and columns.}

\item{ncol}{Number of rows and columns}

\item{scales}{should Scales be fixed (\code{"fixed"}, the default),
free (\code{"free"}), or free in one dimension (\code{"free_x"},
\code{"free_y"}).}

\item{shrink}{If \code{TRUE}, will shrink scales to fit output of
statistics, not raw data. If \code{FALSE}, will be range of raw data
before statistical summary.}

\item{labeller}{A function that takes one data frame of labels and
returns a list or data frame of character vectors. Each input
column corresponds to one factor. Thus there will be more than
one with formulae of the type \code{~cyl + am}. Each output
column gets displayed as one separate line in the strip
label. This function should inherit from the "labeller" S3 class
for compatibility with \code{\link{labeller}()}. See
\code{\link{label_value}} for more details and pointers to other
options.}

\item{as.table}{If \code{TRUE}, the default, the facets are laid out like
a table with highest values at the bottom-right. If \code{FALSE}, the
facets are laid out like a plot with the highest value at the top-right.}

\item{switch}{By default, the labels are displayed on the top and
right of the plot. If \code{"x"}, the top labels will be
displayed to the bottom. If \code{"y"}, the right-hand side
labels will be displayed to the left. Can also be set to
\code{"both"}.}

\item{drop}{If \code{TRUE}, the default, all factor levels not used in the
data will automatically be dropped. If \code{FALSE}, all factor levels
will be shown, regardless of whether or not they appear in the data.}

\item{dir}{Direction: either "h" for horizontal, the default, or "v", for
vertical.}

\item{strip.position}{By default, the labels are displayed on the top of
the plot. Using \code{strip.position} it is possible to place the labels on
either of the four sides by setting \code{strip.position = c("top",
"bottom", "left", "right")}}

\item{page}{The page to draw}
}
\description{
This extension to \code{\link[ggplot2]{facet_wrap}} will allow you to split
a facetted plot over multiple pages. You define a number of rows and columns
per page as well as the page number to plot, and the function will
automatically only plot the correct panels. Usually this will be put in a
loop to render all pages one by one.
}
\note{
If either \code{ncol} or \code{nrow} is \code{NULL} this function will
fall back to the standard \code{facet_wrap} functionality.
}
\examples{
# Calculate the number of pages with 9 panels per page
n_pages <- ceiling(
  length(levels(diamonds$cut)) * length(levels(diamonds$clarity)) / 9
)

# Draw each page
for (i in seq_len(n_pages)) {
  ggplot(diamonds) +
    geom_point(aes(carat, price), alpha = 0.1) +
    facet_wrap_paginate(~cut:clarity, ncol = 3, nrow = 3, page = i)
}

}
\seealso{
Other ggforce.facets: \code{\link{facet_grid_paginate}},
  \code{\link{facet_zoom}}
}

