% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_sets.R
\name{gather_set_data}
\alias{gather_set_data}
\title{Tidy data for use with geom_parallel_sets}
\usage{
gather_set_data(data, x, id_name = "id")
}
\arguments{
\item{data}{A tidy dataframe with some categorical columns}

\item{x}{The columns to use for axes in the parallel sets diagram}

\item{id_name}{The name of the column that will contain the original index of
the row.}
}
\value{
A data.frame
}
\description{
This helper function makes it easy to change tidy data into a tidy(er) format
that can be used by geom_parallel_sets.
}
\examples{
data <- reshape2::melt(Titanic)
head(gather_set_data(data, 1:4))
}
