% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse.R
\name{geom_ellipse}
\alias{geom_ellipse}
\title{Draw ellipses based on the coordinate system scale}
\arguments{
\item{n}{The number of points to sample along the ellipse.}
}
\description{
This is a generalisation of \code{\link{geom_circle}} that allows you to draw
ellipses at a specified angle and center relative to the coordinate system.
}
\section{Aesthetics}{

geom_arc understand the following aesthetics (required aesthetics are in
bold):
\itemize{
 \item{\strong{x0}}
 \item{\strong{y0}}
 \item{\strong{a}}
 \item{\strong{b}}
 \item{\strong{angle}}
 \item{color}
 \item{fill}
 \item{size}
 \item{linetype}
 \item{alpha}
 \item{lineend}
}
}

\section{Computed variables}{


\describe{
 \item{x, y}{The coordinates for the points along the ellipse}
}
}

\examples{
# Basic usage
ggplot() +
  geom_ellipse(aes(x0 = 0, y0 = 0, a = 10, b = 3, angle = 0)) +
  coord_fixed()

# Rotation
# Note that it expects radians and rotates the ellipse counter-clockwise
ggplot() +
  geom_ellipse(aes(x0 = 0, y0 = 0, a = 10, b = 3, angle = pi/4)) +
  coord_fixed()

}
\author{
Thomas Lin Pedersen
}
