% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superformula.R
\name{geom_superformula}
\alias{geom_superformula}
\alias{stat_superformula}
\alias{geom_superformula}
\title{Draw shapes defined by the superformula}
\usage{
stat_superformula(mapping = NULL, data = NULL, geom = "shape",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, n = 100, ...)

geom_superformula(mapping = NULL, data = NULL, stat = "superformula",
  position = "identity", na.rm = FALSE, n = 100, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}} or
\code{\link[=aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[=ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[=fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame.}, and
will be used as the layer data.}

\item{geom}{The geometric object to use display the data}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[=borders]{borders()}}.}

\item{n}{The number of points to calculate around the path}

\item{...}{other arguments passed on to \code{\link[=layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}
}
\description{
The superformula is a generalisation of the formula defining a superellipse
(and, by extension, the formula defining ellipses and circles). By varying
the 7 parameters it allows for description of many shapes found in nature.
}
\section{Aesthetics}{

geom_regon understand the following aesthetics (required aesthetics are in
bold):
\itemize{
\item \strong{x0}
\item \strong{y0}
\item angle
\item a
\item b
\item m1
\item m2
\item n1
\item n2
\item n3
\item revolutions
\item color
\item fill
\item size
\item linetype
\item alpha
\item lineend
}
}

\section{Computed variables}{


\describe{
\item{x, y}{The coordinates for the corners of the polygon}
}
}

\examples{
pars <- expand.grid(m1 = 3:5, n1 = c(0.5, 1, 2, 4))

ggplot(pars) +
    geom_superformula(aes(x0 = 0, y0 = 0, m1 = m1, n1 = n1), n = 1000) +
    facet_grid(m1 ~ n1, labeller = label_both) +
    coord_fixed()


}
\references{
Gielis, J. \emph{A generic geometric transformation that unifies a wide range of natural and abstract shapes.} American Journal of Botany, 90, 333 - 338, (2003).
}
\author{
Thomas Lin Pedersen
}
