% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{interpolateDataFrame}
\alias{interpolateDataFrame}
\title{Interpolate layer data}
\usage{
interpolateDataFrame(data)
}
\arguments{
\item{A}{data.frame with data for a layer}
}
\value{
A similar data.frame with NA values interpolated
}
\description{
Interpolate layer data
}
\keyword{internal}

