% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_circle.R
\name{makeContent.circ_enc}
\alias{makeContent.circ_enc}
\title{Calculate the enclosing circle and draw it as a shapeGrob}
\usage{
\method{makeContent}{circ_enc}(x)
}
\arguments{
\item{x}{A circ_enc grob}
}
\value{
A shape grob
}
\description{
This function takes care of calculating the enclosing circle at draw time and
draw it as a shape grob
}
\keyword{internal}
