% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{makeContent.shape}
\alias{makeContent.shape}
\title{Reactive behaviour for shapeGrob}
\usage{
makeContent.shape(x)
}
\arguments{
\item{x}{A shape grob}
}
\description{
This function defines the way shapeGrob expands and round its polygons
}
\keyword{internal}
