% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{shapeGrob}
\alias{shapeGrob}
\title{The grob powering geom_shape}
\usage{
shapeGrob(
  x = c(0, 0.5, 1, 0.5),
  y = c(0.5, 1, 0.5, 0),
  id = NULL,
  id.lengths = NULL,
  expand = 0,
  radius = 0,
  default.units = "npc",
  name = NULL,
  gp = gpar(),
  vp = NULL
)
}
\arguments{
\item{x}{A numeric vector or unit object specifying x-locations.}

\item{y}{A numeric vector or unit object specifying y-locations.}

\item{id}{A numeric vector used to separate locations in \code{x} and
    \code{y} into multiple polygons.  All locations with the same
    \code{id} belong to the same polygon.}

\item{id.lengths}{A numeric vector used to separate locations in \code{x} and
    \code{y} into multiple polygons.  Specifies consecutive blocks of
    locations which make up separate polygons.}

\item{expand}{An expansion size to expand each shape with, given in units
or a numeric refering to \code{default.units}}

\item{radius}{The corner radius to apply to each shape, given in units
or a numeric refering to \code{default.units}}

\item{default.units}{A string indicating the default units to use
    if \code{x}, \code{y}, \code{width}, or \code{height}
    are only given as numeric vectors.}

\item{name}{ A character identifier. }

\item{gp}{An object of class \code{"gpar"}, typically the output
    from a call to the function \code{\link[grid]{gpar}}.  This is basically
    a list of graphical parameter settings.}

\item{vp}{A Grid viewport object (or NULL).}
}
\value{
A grob of class \code{shape} or, of \code{expand} and \code{radius} are \code{0} a
regular polygon grob
}
\description{
This is the underlying grob constructor for \code{\link[=geom_shape]{geom_shape()}}. It is exported
for others to use but with limited support
}
\keyword{internal}
