% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voronoi.R
\name{geom_voronoi}
\alias{geom_voronoi}
\alias{geom_delaunay}
\alias{geom_voronoi_tile}
\alias{geom_voronoi_segment}
\alias{geom_delaunay_tile}
\alias{geom_delaunay_segment}
\alias{geom_delaunay_segment2}
\alias{stat_delvor_summary}
\title{Voronoi tesselation and delaunay triangulation}
\usage{
geom_voronoi_tile(mapping = NULL, data = NULL, stat = "voronoi_tile",
  position = "identity", na.rm = FALSE, bound = NULL, eps = 1e-09,
  normalize = FALSE, expand = 0, radius = 0, show.legend = NA,
  inherit.aes = TRUE, ...)

geom_voronoi_segment(mapping = NULL, data = NULL,
  stat = "voronoi_segment", position = "identity", na.rm = FALSE,
  bound = NULL, eps = 1e-09, normalize = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)

geom_delaunay_tile(mapping = NULL, data = NULL, stat = "delaunay_tile",
  position = "identity", na.rm = FALSE, bound = NULL, eps = 1e-09,
  normalize = FALSE, expand = 0, radius = 0, show.legend = NA,
  inherit.aes = TRUE, ...)

geom_delaunay_segment(mapping = NULL, data = NULL,
  stat = "delaunay_segment", position = "identity", na.rm = FALSE,
  bound = NULL, eps = 1e-09, normalize = normalize, show.legend = NA,
  inherit.aes = TRUE, ...)

geom_delaunay_segment2(mapping = NULL, data = NULL,
  stat = "delaunay_segment2", position = "identity", na.rm = FALSE,
  bound = NULL, eps = 1e-09, normalize = FALSE, n = 100,
  show.legend = NA, inherit.aes = TRUE, ...)

stat_delvor_summary(mapping = NULL, data = NULL, geom = "point",
  position = "identity", na.rm = FALSE, bound = NULL, eps = 1e-09,
  normalize = FALSE, show.legend = NA, inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link{aes}} or
\code{\link{aes_}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
   options:

   If \code{NULL}, the default, the data is inherited from the plot
   data as specified in the call to \code{\link{ggplot}}.

   A \code{data.frame}, or other object, will override the plot
   data. All objects will be fortified to produce a data frame. See
   \code{\link{fortify}} for which variables will be created.

   A \code{function} will be called with a single argument,
   the plot data. The return value must be a \code{data.frame.}, and
   will be used as the layer data.}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{bound}{The bounding rectangle for the tesselation. Defaults to
\code{NULL} which creates a rectangle expanded 10\% in all directions. If
supplied it should be a vector giving the bounds in the following order:
xmin, xmax, ymin, ymax.}

\item{eps}{A value of epsilon used in testing whether a quantity is zero,
mainly in the context of whether points are collinear. If anomalous errors
arise, it is possible that these may averted by adjusting the value of eps
upward or downward.}

\item{normalize}{Should coordinates be normalized prior to calculations. If
\code{x} and \code{y} are in wildly different ranges it can lead to
tesselation and triangulation that seems off when plotted without
\code{\link[ggplot2]{coord_fixed}}. Normalization of coordinates solves this.
The coordinates are transformed back after calculations.}

\item{expand}{A numeric or unit vector of length one, specifying the
expansion amount. Negative values will result in contraction instead. If the
value is given as a numeric it will be understood as a proportion of the
plot area width.}

\item{radius}{As \code{expand} but specifying the corner radius.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link{borders}}.}

\item{...}{other arguments passed on to \code{\link{layer}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{n}{The number of points to create for each segment}

\item{geom}{stat Override the default connection between \code{geom_arc} and
\code{stat_arc}.}
}
\description{
This set of geoms and stats allows you to display voronoi tesselation and
delaunay triangulation, both as polygons and as line segments. Furthermore
it lets you augment your point data with related summary statistics. The
computations are based on the \code{\link[deldir]{deldir}} package.
}
\section{Aesthetics}{

geom_voronoi_tile and geom_delaunay_tile understand the following aesthetics
(required aesthetics are in bold):
\itemize{
 \item{\strong{x}}
 \item{\strong{y}}
 \item{alpha}
 \item{color}
 \item{fill}
 \item{linetype}
 \item{size}
}
geom_voronoi_segment, geom_delaunay_segment, and geom_delaunay_segment2
understand the following aesthetics (required aesthetics are in bold):
\itemize{
 \item{\strong{x}}
 \item{\strong{y}}
 \item{alpha}
 \item{color}
 \item{linetype}
 \item{size}
}
}

\section{Computed variables}{

stat_delvor_summary computes the following variables:
\describe{
 \item{x, y}{If \code{switch.centroid = TRUE} this will be the coordinates for
 the voronoi tile centroid, otherwise it is the original point}
 \item{xcent, ycent}{If \code{switch.centroid = FALSE} this will be the
 coordinates for the voronoi tile centroid, otherwise it will be \code{NULL}}
 \item{xorig, yorig}{If \code{switch.centroid = TRUE} this will be the
 coordinates for the original point, otherwise it will be \code{NULL}}
 \item{ntri}{Number of triangles emanating from the point}
 \item{triarea}{The total area of triangles emanating from the point divided
 by 3}
 \item{triprop}{\code{triarea} divided by the sum of the area of all
 triangles}
 \item{nsides}{Number of sides on the voronoi tile associated with the point}
 \item{nedges}{Number of sides of the associated voronoi tile that is part of
 the bounding box}
 \item{vorarea}{The area of the voronoi tile associated with the point}
 \item{vorprop}{\code{vorarea} divided by the sum of all voronoi tiles}
}
}

\examples{
# Voronoi
ggplot(iris, aes(Sepal.Length, Sepal.Width)) +
  geom_voronoi_tile(aes(fill = Species)) +
  geom_voronoi_segment() +
  geom_text(aes(label = ..nsides.., size = ..vorarea..),
            stat = 'delvor_summary', switch.centroid = TRUE)

# Difference of normalize = TRUE
ggplot(iris, aes(Sepal.Length, Sepal.Width)) +
  geom_voronoi_tile(aes(fill = Species), normalize = TRUE) +
  geom_voronoi_segment(normalize = TRUE) +
  geom_text(aes(label = ..nsides.., size = ..vorarea..),
            stat = 'delvor_summary', switch.centroid = TRUE, normalize = TRUE)

# Delaunay triangles - interpolation of species
ggplot(iris, aes(Sepal.Length, Sepal.Width)) +
  geom_delaunay_tile(alpha = 0.3) +
  geom_delaunay_segment2(aes(colour = Species, group = -1))

}
\author{
Thomas Lin Pedersen
}
