% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc_bar.R, R/arc.R, R/bezier.R,
%   R/bspline.R, R/circle.R, R/facet_grid_paginate.R, R/facet_stereo.R,
%   R/facet_wrap_paginate.R, R/facet_zoom.R, R/ggproto-classes.R,
%   R/interpolate.R, R/link.R, R/scale-unit.R, R/sina.R, R/voronoi.R
\docType{data}
\name{StatArcBar}
\alias{StatArcBar}
\alias{StatPie}
\alias{GeomArcBar}
\alias{StatArc}
\alias{GeomArc}
\alias{StatArc2}
\alias{StatArc0}
\alias{GeomArc0}
\alias{StatBezier}
\alias{StatBezier2}
\alias{StatBezier0}
\alias{GeomBezier0}
\alias{StatBspline}
\alias{StatBspline2}
\alias{GeomBspline0}
\alias{StatCircle}
\alias{GeomCircle}
\alias{FacetGridPaginate}
\alias{FacetStereo}
\alias{FacetWrapPaginate}
\alias{FacetZoom}
\alias{ggforce-extensions}
\alias{GeomPathInterpolate}
\alias{StatLink}
\alias{StatLink2}
\alias{ScaleContinuousPositionUnit}
\alias{StatSina}
\alias{GeomSina}
\alias{StatVoronoiTile}
\alias{StatVoronoiSegment}
\alias{StatDelaunayTile}
\alias{StatDelaunaySegment}
\alias{StatDelaunaySegment2}
\alias{StatDelvorSummary}
\title{ggforce extensions to ggplot2}
\description{
ggforce makes heavy use of the ggproto class system to extend the
functionality of ggplot2. In general the actual classes should be of little
interest to users as the standard ggplot2 api of using geom_* and stat_*
functions for building up the plot is encouraged.
}
\keyword{datasets}
