% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans.R
\name{power_trans}
\alias{power_trans}
\title{Create a power transformation object}
\usage{
power_trans(n)
}
\arguments{
\item{n}{The degree of the power transformation}
}
\value{
A trans object
}
\description{
This function can be used to create a proper trans object that encapsulates
a power transformation (x^n).
}

