% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-unit.R
\name{scale_unit}
\alias{scale_x_unit}
\alias{scale_y_unit}
\title{Position scales for units data}
\usage{
scale_x_unit(...)

scale_y_unit(...)
}
\arguments{
\item{...}{Passed on to \code{units::scale_x_units()} or \code{units::scale_y_units()}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} These are the default scales for the units
class. These will usually be added automatically. To override manually, use
\verb{scale_*_unit}.
}
\keyword{internal}
